+++ broqer/op/combine_latest.py
@@ -73,14 +73,14 @@ class CombineLatest(MultiOperator):
     def unsubscribe(self, subscriber: Subscriber) -> None:
         MultiOperator.unsubscribe(self, subscriber)
         if not self._subscriptions:
-            self._missing = set(self._orginators)
+            self._missing = set(self._originators)
             self._partial_state[:] = [NONE for _ in self._partial_state]
 
     def get(self):
         if self._subscriptions:
             return self._state
 
-        values = tuple(p.get() for p in self._orginators)
+        values = tuple(p.get() for p in self._originators)
 
         if NONE in values:
             return NONE
@@ -91,7 +91,7 @@ class CombineLatest(MultiOperator):
         return self._map(*values)
 
     def emit(self, value: Any, who: Publisher) -> None:
-        if all(who is not p for p in self._orginators):
+        if all(who is not p for p in self._originators):
             raise ValueError('Emit from non assigned publisher')
 
         # remove source publisher from ._missing